Internal #861: Refactor WindowConstantAggregator
Move the WindowConstantAggregator class to separate files.
