+++ src/function/window/CMakeLists.txt
@@ -6,6 +6,7 @@ add_library_unity(
   window_aggregator.cpp
   window_boundaries_state.cpp
   window_collection.cpp
+  window_constant_aggregator.cpp
   window_cumedist_function.cpp
   window_executor.cpp
   window_naive_aggregator.cpp
